Transform remapped values to the correct type before comparison
In some databases (i.e. Oracle) query responses can contain different
datatypes than what the JSON can support (the encoding for the values
column of the field values table). This commit adds a funciton that
transforms the values from the field values table to the format the
database will return.